+++ backend_api/Controllers/StockController.cs
@@ -6,6 +6,7 @@ using backend_api.Data;
 using backend_api.Dtos.Stock;
 using backend_api.Interfaces;
 using backend_api.Mappers;
+using backend_api.Helpers;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -27,9 +28,11 @@ namespace backend_api.Controllers  // 确保这是正确的命名空间
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        public async Task<IActionResult> GetAll([FromQuery] QueryObject query)
         {
-            var stocks = await _stockRepository.GetAllAsync();
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var stocks = await _stockRepository.GetAllAsync(query);
 
             var stockDto = stocks.Select(s => s.ToStockDto());
 
@@ -38,9 +41,11 @@ namespace backend_api.Controllers  // 确保这是正确的命名空间
         }
 
         // Database 才需要 await
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);           
             var stock = await _stockRepository.GetByIdAsync(id);
             if (stock == null)
             {
@@ -60,6 +65,8 @@ namespace backend_api.Controllers  // 确保这是正确的命名空间
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);           
             var stockModel = stockDto.ToStockFromCreateDTO();
             stockModel = await _stockRepository.CreateAsync(stockModel);
             return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
@@ -67,9 +74,11 @@ namespace backend_api.Controllers  // 确保这是正确的命名空间
         
         // Database 需要 await
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);           
             var stockModel = await _stockRepository.UpdateAsync(id, stockDto);
             if (stockModel == null)
             {
@@ -79,9 +88,11 @@ namespace backend_api.Controllers  // 确保这是正确的命名空间
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);           
             var stockModel = await _stockRepository.DeleteAsync(id);
             if (stockModel == null)
             {
